+++ pyo3-derive-backend/src/lib.rs
@@ -5,6 +5,7 @@
 
 mod defs;
 mod func;
+mod konst;
 mod method;
 mod module;
 mod pyclass;